requirements.txt
Flask==3.0.0
pymongo==4.6.1
flask-login==0.6.3
werkzeug==3.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
dnspython==2.4.2
